requirements.txt
alembic==1.14.1
aniso8601==10.0.1
asttokens==3.0.0
backcall==0.2.0
click==8.1.3
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
Faker==18.9.0
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pluggy==1.5.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
SQLAlchemy-serializer==1.4.1
stack-data==0.6.3
tabulate==0.9.0
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.13
Werkzeug==2.2.2
zipp==3.20.2
